added pkcs8 support and made a code clean up
