+++ lib/src/authentication/cryptosign/ppk_mac_data.dart
@@ -1,0 +1,9 @@
+import 'dart:typed_data';
+
+class PpkMacData {
+  late String algorithm;
+  late String encryption;
+  late String comment;
+  late Uint8List publicKey;
+  late Uint8List privateKey;
+}
